CDMS-476: Enable validation on ClearanceRequest, Finalisation and ServiceHeaders (#58)
* CDMS-476: Enable validation on ClearanceRequest, Finalisation and ServiceHeaders

This enables the validation with the currently implemented validators.
At the moment it will just skip the message, which is the intended behaviour, but only logs the errors out and does not send them onwards.

The fixture generation functions required some more explicit configuration to ensure that they were valid for the rules.

A couple of fixes were identified for the validation rules which have been updated.

* CDMS-476: Add docker compose logs to the pipeline

* CDMS-476: Add message type to the validation output logs

* CDMS-746: Update InboundError fixture to use the ServiceHeader fixture

* CDMS-476: Fix test set up assertion