CDMS-649: Optimise logic around new notification validation (#85)
There is a potential that a notification that has a status of VALIDATED can come in microseconds after a notification is stored that is IN_PROGRESS.
This currently results in the message being skipped.

This updates the logic to check whether the new notification is later in the lifecycle of a notification to aid in deciding whether to skip the message.