CDMS-649: Refine message rejection logic (#113)
Messages can arrive within milliseconds of each other which has the potential for messages to be skipped when inspecting the timestamp.
This changes the logic to only check the timestamp if it is the same status or always progress if there is a change to the status and it is more mature notification.