CDMS-968: Update ALVSVAL403 with new logic (#209)
* CDMS-968: Update ALVSVAL403 with new logic

We should only raise ALVSVAL403 if a non-cancellation finalisation is processed where a cancelled finalisation already exists.
In other words, the new finalisation is trying to uncancel the existing one.

* CDMS-968: Fix an issue with ALVSVAL506

This was triggering when we received a finalisation through which was not a cancellation and we had previously received a finalisation.
We shouldn't be receiving more than one finalisation so this shouldn't actually occur, but it is now fixed anyway.